pipebox: rename `CliftContainer.h`
The new name is `CliftContainers.h`.
